Now tag type lazily evaluates existing tags ...
* note to self: do not initiate a list from a lazy collection (ormlite)
	* using .iterator() is the best way to go with it

+++ src/main/java/edu/brandeis/cs/nlp/mae/database/ExtentTagDao.java
@@ -105,11 +105,15 @@ public class ExtentTagDao extends BaseDaoImpl<ExtentTag, String> {
         final Collection<Attribute> atts = tag.getAttributes();
         final Collection<CharIndex> anchors = tag.getSpans();
         super.callBatchTasks((Callable<Void>) () -> {
-            for (Attribute att : atts) {
-                attDao.delete(att);
+            if (atts != null) {
+                for (Attribute att : atts) {
+                    attDao.delete(att);
+                }
             }
-            for (CharIndex anchor : anchors) {
-                charIndexDao.delete(anchor);
+            if (anchors != null) {
+                for (CharIndex anchor : anchors) {
+                    charIndexDao.delete(anchor);
+                }
             }
             return null;
         });
